requirements.txt
anyio==3.5.0; python_full_version >= "3.6.2" and python_version >= "3.7"
astroid==2.11.2; python_full_version >= "3.6.2"
async-generator==1.10; python_version >= "3.7"
atomicwrites==1.4.0; python_version >= "3.7" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.7" and python_full_version >= "3.4.0"
attrs==21.4.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
black==22.3.0; python_full_version >= "3.6.2"
certifi==2021.10.8; python_version >= "3.7"
cffi==1.15.0; os_name == "nt" and implementation_name != "pypy" and python_version >= "3.7"
cfgv==3.3.1; python_full_version >= "3.6.1" and python_version >= "3.7"
charset-normalizer==2.0.12; python_full_version >= "3.5.0" and python_version >= "3.7"
click==8.1.2; python_version >= "3.7" and python_full_version >= "3.6.2"
colorama==0.4.4; sys_platform == "win32" and python_version >= "3.7" and python_full_version >= "3.6.2" and (python_version >= "3.7" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.7" and python_full_version >= "3.5.0") and platform_system == "Windows"
coverage==6.3.2; python_version >= "3.7"
dill==0.3.4; python_full_version >= "3.6.2"
distlib==0.3.4; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
filelock==3.6.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
h11==0.12.0; python_version >= "3.6"
httpcore==0.14.7; python_version >= "3.7"
httpx-caching==0.1a2; python_version >= "3.7"
httpx==0.22.0; python_version >= "3.6"
identify==2.4.12; python_version >= "3.7"
idna==3.3; python_version >= "3.7" and python_full_version >= "3.6.2"
iniconfig==1.1.1; python_version >= "3.7"
isort==5.10.1; python_full_version >= "3.6.1" and python_version < "4.0"
lazy-object-proxy==1.7.1; python_version >= "3.6" and python_full_version >= "3.6.2"
loguru==0.6.0; python_version >= "3.5"
mccabe==0.7.0; python_version >= "3.6" and python_full_version >= "3.6.2"
msgpack==1.0.3; python_version >= "3.7"
multimethod==1.8; python_version >= "3.7"
mypy-extensions==0.4.3; python_full_version >= "3.6.2" and python_version >= "3.6"
mypy==0.942; python_version >= "3.6"
nodeenv==1.6.0; python_version >= "3.7"
outcome==1.1.0; python_version >= "3.7"
packaging==21.3; python_version >= "3.7"
pathspec==0.9.0; python_full_version >= "3.6.2"
pillow==9.1.0; python_version >= "3.7"
platformdirs==2.5.1; python_version >= "3.7" and python_full_version >= "3.6.2" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7")
pluggy==1.0.0; python_version >= "3.7"
pre-commit==2.18.1; python_version >= "3.7"
py==1.11.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
pycparser==2.21; python_version >= "3.7" and python_full_version < "3.0.0" and os_name == "nt" and implementation_name != "pypy" or os_name == "nt" and implementation_name != "pypy" and python_version >= "3.7" and python_full_version >= "3.4.0"
pylint==2.13.5; python_full_version >= "3.6.2"
pyparsing==3.0.8; python_full_version >= "3.6.8" and python_version >= "3.7"
pytest-cov==3.0.0; python_version >= "3.6"
pytest-httpx==0.20.0; python_version >= "3.6"
pytest-trio==0.7.0; python_version >= "3.6"
pytest==7.1.1; python_version >= "3.7"
pyyaml==6.0; python_version >= "3.7"
rfc3986==1.5.0; python_version >= "3.7"
six==1.16.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
sniffio==1.2.0; python_full_version >= "3.6.2" and python_version >= "3.7"
sortedcontainers==2.4.0; python_version >= "3.7"
toml==0.10.2; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.7"
tomli==2.0.1; python_version < "3.11" and python_full_version >= "3.6.2" and python_version >= "3.7"
trio==0.20.0; python_version >= "3.7"
typing-extensions==4.1.1; python_version >= "3.6"
ujson==5.2.0; python_version >= "3.7"
virtualenv==20.14.1; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
win32-setctime==1.1.0; sys_platform == "win32" and python_version >= "3.5"
wrapt==1.14.0; python_full_version >= "3.6.2"
